Add test-output to RAT exclusion list
There is an inconsistency as to where test-output reports are
created.  In Eclipse, they are created outside of the target
directory when running unit tests with TestNG.  However, when
using Surefire and Maven the test-output is created in the
correct location.  The safest course of action is to exclude
this folder from RAT checks.