For all public classes in /memory/ (and not /memory/internal/):
1. I have corrected the javadocs and code comments to use "off-heap"
instead of "direct" or "native". However, I decided not to change the
use of "Direct" in existing method calls or in the /internal/ directory.
That would cause a huge number of changes (100's).

2. I have corrected the javadocs and code comments to use "alive" or
"isAlive" instead of "valid" or "isValid". However, I decided not to
change the word "Valid" in the /internal/ directory.  That would cause a
huge number of changes (100's).